+++ ui/packages/shared/profile/src/ProfileView/hooks/useResetStateOnProfileTypeChange.ts
@@ -1,0 +1,32 @@
+// Copyright 2022 The Parca Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import {useURLState} from '@parca/components';
+
+export const useResetStateOnProfileTypeChange = (): (() => void) => {
+  const [, setGroupBy] = useURLState('group_by');
+  const [, setFilterByFunction] = useURLState('filter_by_function');
+  const [, setExcludeFunction] = useURLState('exclude_function');
+  const [, setSearchString] = useURLState('search_string');
+  const [, setCurPath] = useURLState('cur_path');
+
+  return () => {
+    setTimeout(() => {
+      setGroupBy(undefined);
+      setFilterByFunction(undefined);
+      setExcludeFunction(undefined);
+      setSearchString(undefined);
+      setCurPath(undefined);
+    });
+  };
+};