+++ ui/packages/shared/components/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.344](https://github.com/parca-dev/parca/compare/@parca/components@0.16.343...@parca/components@0.16.344) (2025-06-20)
+
+**Note:** Version bump only for package @parca/components
+
+## 0.16.343 (2025-06-19)
+
+**Note:** Version bump only for package @parca/components
+
+## 0.16.342 (2025-06-18)
+
+**Note:** Version bump only for package @parca/components
+
 ## [0.16.341](https://github.com/parca-dev/parca/compare/@parca/components@0.16.340...@parca/components@0.16.341) (2025-06-16)
 
 **Note:** Version bump only for package @parca/components