+++ ui/packages/shared/profile/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.19.6](https://github.com/parca-dev/parca/compare/@parca/profile@0.19.5...@parca/profile@0.19.6) (2025-06-23)
+
+**Note:** Version bump only for package @parca/profile
+
+## [0.19.5](https://github.com/parca-dev/parca/compare/@parca/profile@0.19.4...@parca/profile@0.19.5) (2025-06-20)
+
+**Note:** Version bump only for package @parca/profile
+
+## [0.19.4](https://github.com/parca-dev/parca/compare/@parca/profile@0.19.3...@parca/profile@0.19.4) (2025-06-19)
+
+**Note:** Version bump only for package @parca/profile
+
+## [0.19.3](https://github.com/parca-dev/parca/compare/@parca/profile@0.19.2...@parca/profile@0.19.3) (2025-06-19)
+
+**Note:** Version bump only for package @parca/profile
+
+## 0.19.2 (2025-06-19)
+
+**Note:** Version bump only for package @parca/profile
+
+## [0.19.1](https://github.com/parca-dev/parca/compare/@parca/profile@0.19.0...@parca/profile@0.19.1) (2025-06-18)
+
+**Note:** Version bump only for package @parca/profile
+
+# [0.19.0](https://github.com/parca-dev/parca/compare/@parca/profile@0.18.4...@parca/profile@0.19.0) (2025-06-18)
+
+### Features
+
+- implement sandwich view ([#5683](https://github.com/parca-dev/parca/issues/5683)) ([83b54de](https://github.com/parca-dev/parca/commit/83b54ded3dc6626889421689d0293bda6d199fbd))
+
+## [0.18.4](https://github.com/parca-dev/parca/compare/@parca/profile@0.18.3...@parca/profile@0.18.4) (2025-06-18)
+
+**Note:** Version bump only for package @parca/profile
+
+## [0.18.3](https://github.com/parca-dev/parca/compare/@parca/profile@0.18.2...@parca/profile@0.18.3) (2025-06-18)
+
+**Note:** Version bump only for package @parca/profile
+
+## [0.18.2](https://github.com/parca-dev/parca/compare/@parca/profile@0.18.1...@parca/profile@0.18.2) (2025-06-17)
+
+**Note:** Version bump only for package @parca/profile
+
+## [0.18.1](https://github.com/parca-dev/parca/compare/@parca/profile@0.18.0...@parca/profile@0.18.1) (2025-06-17)
+
+**Note:** Version bump only for package @parca/profile
+
 # [0.18.0](https://github.com/parca-dev/parca/compare/@parca/profile@0.17.11...@parca/profile@0.18.0) (2025-06-16)
 
 ### Features